move more utils to tests/base; delete unused memory.py
