+++ tests/contrib/httplib/test_httplib.py
@@ -16,9 +16,7 @@ from ddtrace.pin import Pin
 
 from tests.opentracing.utils import init_tracer
 
-from ...base import BaseTracerTestCase
-from ...util import override_global_tracer
-from tests.base import assert_span_http_status_code
+from tests.base import assert_span_http_status_code, BaseTracerTestCase, override_global_tracer
 
 if PY2:
     from urllib2 import urlopen, build_opener, Request
